fix[schema, database-client]: remove ceramic client, fix faiilng tests
