+++ database-client/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./ceramicClient";
+export * from "./logger";
 export * from "./passportScorerClient";
 export * from "./composeDatabase";
 export * from "./types";
